Return pairwise predictions for specified agents

Given a mELO_rating model and a vector of agents, this function gives the predicted success probabilities for each combination. This is useful for mELO models as the ratings alone is not useful for predicting outcomes.

model_pred_mat(model, agents, round = 4)

Arguments

model
An object of class mELO_rating.

agents
a vector of agents/players/teams you would like to pairwise predictions for.

round
Round results to how many figures. Default is 4.

actual_matrix
Should result be an actual matrix? Default is FALSE and the output is a data.frame.

Value

a data.frame.

Examples

# Fit a mELO model that can handle these types of interactions.
rps_mELO <- mELO(rps_df, k=1)
rps_mELO

#> 
#> mELO Ratings For 3 Players Playing 120 Games
#> 
#> k = 1.
#> 
#>     Player Rating Games Win Draw Loss Lag
#> 1    PAPER 2205.8    80  40    0   40   0
#> 2 SCISSORS 2204.7    80  40    0   40   0
#> 3     ROCK 2189.5    80  40    0   40   1
#> 

# Get pairwise predictions
model_pred_mat(
   rps_mELO,
   unique(rps_df[[2]])
)

#>           PAPER   ROCK SCISSORS
#> PAPER    0.5000 0.9986   0.0014
#> ROCK     0.0014 0.5000   0.9987
#> SCISSORS 0.9986 0.0013   0.5000